requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
pycparser==2.19
python-telegram-bot==12.0.0b1
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.3
